requirements.txt
blinker==1.6.2
cachelib==0.10.2
click==8.1.6
Flask==2.2.5
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.5
gunicorn==21.2.0
numpy==1.25.1
pandas==2.0.3
python-http-client==3.3.7
starkbank-ecdsa==2.2.0
PyJWT==2.8.0
requests==2.31.0
livereload==2.6.3
Werkzeug==2.2.2
